BT-29 validation des donnees (date,string et double)

+++ src/ui/ClientUi.java
@@ -4,6 +4,7 @@ import entity.Client;
 import repository.impl.ClientRepositoryImpl;
 import service.ClientService;
 import service.impl.ClientServiceImpl;
+import utils.InputValidator;
 
 import java.util.List;
 import java.util.Optional;
@@ -51,16 +52,28 @@ public class ClientUi {
         }
     }
 
-    private void createClient() {
+    public void createClient() {
         System.out.println("Ajouter un client");
         System.out.println("Veullez entrer le nom du client");
         String name = scanner.nextLine();
+        while (!InputValidator.validateString(name)) {
+            System.out.println("Erreur : Nom incorrect. Veuillez réessayer.");
+            name = scanner.next();
+        }
 
         System.out.println("Veuillez entrer l'adresse du client'");
         String adresse = scanner.nextLine();
+        while (!InputValidator.validateString(adresse)) {
+            System.out.println("Erreur : Adreesee incorrect. Veuillez réessayer.");
+            adresse = scanner.next();
+        }
 
         System.out.println("Veuillez entrer le phonenUmber du client");
         String phoneNumber = scanner.nextLine();
+        while (!InputValidator.validateString(phoneNumber)) {
+            System.out.println("Erreur : Numero de telephone incorrect. Veuillez réessayer.");
+            phoneNumber = scanner.next();
+        }
 
         System.out.println("Le client est-il professionnel (1 : oui, 0 : non) ? ");
         int estProf = scanner.nextInt();
@@ -78,7 +91,6 @@ public class ClientUi {
 
         Client client = new Client(name, adresse, phoneNumber, professionnel);
         clientService.createClient(client);
-        menu();
     }
 
     private void showAllClients() {
@@ -115,33 +127,51 @@ public class ClientUi {
         Long id = scanner.nextLong();
         scanner.nextLine();
 
-        System.out.println("Veullez entrer le nouveau nom");
-        String name = scanner.nextLine();
 
-        System.out.println("Veuillez entrer la nouvelle adresse'");
-        String adresse = scanner.nextLine();
+        Optional<Client> optionalClient = clientService.getClientById(id);
 
-        System.out.println("Veuillez entrer le nouveau phonenUmber");
-        String phoneNumber = scanner.nextLine();
+        if (optionalClient.isPresent()) {
+            System.out.println("Veullez entrer le nouveau nom");
+            String name = scanner.nextLine();
+            while (!InputValidator.validateString(name)) {
+                System.out.println("Erreur : Nom non valide. Veuillez réessayer.");
+                name = scanner.next();
+            }
 
-        System.out.println("Le client est-il professionnel (1 : oui, 0 : non) ? ");
-        int estProf = scanner.nextInt();
-        scanner.nextLine();
+            System.out.println("Veuillez entrer la nouvelle adresse'");
+            String adresse = scanner.nextLine();
+            while (!InputValidator.validateString(adresse)) {
+                System.out.println("Erreur : Adreesee incorrect. Veuillez réessayer.");
+                adresse = scanner.next();
+            }
 
-        boolean professionnel;
-        if (estProf == 1) {
-            professionnel = true;
-        } else if (estProf == 0) {
-            professionnel = false;
+            System.out.println("Veuillez entrer le nouveau phonenUmber");
+            String phoneNumber = scanner.nextLine();
+            while (!InputValidator.validateString(phoneNumber)) {
+                System.out.println("Erreur : Numero de telephone incorrect. Veuillez réessayer.");
+                phoneNumber = scanner.next();
+            }
+
+            System.out.println("Le client est-il professionnel (1 : oui, 0 : non) ? ");
+            int estProf = scanner.nextInt();
+            scanner.nextLine();
+
+            boolean professionnel;
+            if (estProf == 1) {
+                professionnel = true;
+            } else if (estProf == 0) {
+                professionnel = false;
+            } else {
+                System.out.println("Erreur: choix de disponibilité invalide.");
+                return;
+            }
+
+            Client client = new Client(name, adresse, phoneNumber, professionnel);
+            clientService.updateClient(client, id);
+
         } else {
-            System.out.println("Erreur: choix de disponibilité invalide.");
-            return;
+            System.out.println("Client non trouvé avec l'id " + id);
         }
-
-        Client client = new Client(name, adresse, phoneNumber, professionnel);
-        clientService.updateClient(client, id);
-
-        menu();
     }
 
     private void deleteClient() {
@@ -149,8 +179,13 @@ public class ClientUi {
         Long id = scanner.nextLong();
         scanner.nextLine();
 
+        Optional<Client> optionalClient = clientService.getClientById(id);
+
+        if (optionalClient.isPresent()) {
         clientService.deleteClient(id);
-        menu();
+        } else {
+            System.out.println("Client non trouvé avec l'id " + id);
+        }
     }
 
 }
